Unit_4/u4d7/django/one_project/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
django==4.0.4; python_version >= '3.8'
djangorestframework-simplejwt==5.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
psycopg==3.0.11
pyjwt==2.3.0; python_version >= '3.6'
pytz==2022.1
setuptools==62.1.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
